+++ praudio/errors.py
@@ -2,3 +2,9 @@
 
 class FileExtensionError(Exception):
     """Error that is thrown when the extension of a file isn't allowed."""
+
+class NotA2DSignalError(Exception):
+    """
+    Error that is thrown when the expected signal should have more than 1
+    dimension
+    """