+++ test/transform/utils.py
@@ -9,3 +9,9 @@ from praudio.io.signal import Signal
 def sample_signal():
     array = np.random.rand(6000)
     return Signal("dummy", 22050, array, "dummy/file/wav")
+
+
+@pytest.fixture
+def sample_2d_signal():
+    array = np.random.rand(64, 200)
+    return Signal("dummy", 22050, array, "dummy/file/wav")